third_party/WebKit/LayoutTests/fast/dom/Window/window-frames-self-referential.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>

<script>
description('This tests that window.frames is self-referential.');

shouldBe("window.frames", "window");
</script>
</body>
</html>
